refactor: web auth controller
