+++ packages/totem_app/lib/features/auth/screens/pin_entry_screen.dart
@@ -3,6 +3,7 @@ import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:go_router/go_router.dart';
 import 'package:pinput/pinput.dart';
+import 'package:totem_app/features/auth/controllers/auth_controller.dart';
 import 'package:totem_core/auth/controllers/auth_controller.dart';
 import 'package:totem_core/auth/models/auth_state.dart';
 import 'package:totem_core/core/config/app_config.dart';
@@ -61,7 +62,7 @@ class _PinEntryScreenState extends ConsumerState<PinEntryScreen> {
 
     try {
       await ref
-          .read(authControllerProvider.notifier)
+          .read(mobileAuthControllerProvider)
           .verifyPin(_pinController.text.trim());
       Future<void>.delayed(const Duration(seconds: 2), () {
         if (mounted) {
