+++ packages/totem_app/lib/features/auth/screens/profile_setup_screen.dart
@@ -3,7 +3,7 @@ import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:go_router/go_router.dart';
-import 'package:totem_core/auth/controllers/auth_controller.dart';
+import 'package:totem_app/features/auth/controllers/auth_controller.dart';
 import 'package:totem_core/core/api/api_client/api_client.dart';
 import 'package:totem_core/core/config/app_config.dart';
 import 'package:totem_core/core/config/theme.dart';
@@ -71,7 +71,7 @@ class _ProfileSetupScreenState extends ConsumerState<ProfileSetupScreen> {
     setState(() => _isLoading = true);
     try {
       await ref
-          .read(authControllerProvider.notifier)
+          .read(mobileAuthControllerProvider)
           .completeOnboarding(
             firstName: _firstNameController.text.trim(),
             referralSource: _referralSource,
