+++ packages/totem_app/lib/features/profile/screens/profile_details_screen.dart
@@ -1,5 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
+import 'package:totem_app/features/auth/controllers/auth_controller.dart';
 import 'package:totem_core/auth/controllers/auth_controller.dart';
 import 'package:totem_core/core/config/theme.dart';
 import 'package:totem_core/core/errors/error_handler.dart';
@@ -45,7 +46,7 @@ class _ProfileDetailsScreenState extends ConsumerState<ProfileDetailsScreen> {
     });
 
     final updated = await ref
-        .read(authControllerProvider.notifier)
+        .read(mobileAuthControllerProvider)
         .updateUserProfile(
           name: _nameController.text.trim(),
           email: _emailController.text.trim(),
